+++ doomclassic/doom/r_main.cpp
@@ -417,7 +417,7 @@ void R_InitPointToAngle (void)
 	// UNUSED - now getting from tables.c
 #if 0
 	int	i;
-	long	t;
+    int    t;
 	float	f;
 	//
 	// slope (tangent) to angle lookup
