+++ neo/sys/sys_net.cpp
@@ -105,8 +105,8 @@ static SOCKET	socks_socket;
 static char		socksBuf[4096];
 
 typedef struct {
-	unsigned long ip;
-	unsigned long mask;
+	unsigned int ip;
+	unsigned int mask;
 	char addr[16];
 } net_interface;
 
@@ -261,7 +261,7 @@ static bool Net_StringToSockaddr( const char *s, sockaddr_in *sadr, bool doDNSRe
 	sadr->sin_port = 0;
 
 	if( s[0] >= '0' && s[0] <= '9' ) {
-		unsigned long ret = inet_addr(s);
+		unsigned int ret = inet_addr(s);
 		if ( ret != INADDR_NONE ) {
 			*(int *)&sadr->sin_addr = ret;
 		} else {
@@ -300,7 +300,7 @@ NET_IPSocket
 int NET_IPSocket( const char *net_interface, int port, netadr_t *bound_to ) {
 	SOCKET				newsocket;
 	sockaddr_in			address;
-	unsigned long		_true = 1;
+	unsigned int        _true = 1;
 	int					i = 1;
 	int					err;
 
@@ -727,7 +727,7 @@ void Sys_InitNetworking() {
 			idLib::Printf( "Found interface: %s %s - ", pAdapter->AdapterName, pAdapter->Description );
 			pIPAddrString = &pAdapter->IpAddressList;
 			while( pIPAddrString ) {
-				unsigned long ip_a, ip_m;
+				unsigned int ip_a, ip_m;
 				if( !idStr::Icmp( "127.0.0.1", pIPAddrString->IpAddress.String ) ) {
 					foundloopback = true;
 				}
@@ -879,9 +879,9 @@ bool Sys_IsLANAddress( const netadr_t adr ) {
 
 	if ( num_interfaces ) {
 		int i;
-		unsigned long *p_ip;
-		unsigned long ip;
-		p_ip = (unsigned long *)&adr.ip[0];
+		unsigned int *p_ip;
+		unsigned int ip;
+		p_ip = (unsigned int *)&adr.ip[0];
 		ip = ntohl( *p_ip );
 
 		for( i = 0; i < num_interfaces; i++ ) {
